Started creating proper unit tests
The class TestingRelayCat emulates an irc connection without
connecting to a remote server (although, a PircBot instance is initialised).
It allows the tests to emulate being remote users, sending messages
both directly and via channels to RelayCat.
The relpies are stored in a queue, which the tests can examine for
correctness.

This has it's one set of unit tests, in TestingRelayCatTest, which
also demonstrates some of the usage.