Remove LocalBufferedSocket close(), add note to flush()
close() was added as a way of getting round some really silly
issues with buffered readers and writers; by adding an end of stream
data point, the buffer was not marked as under-filled, and things work.
Later investiagation showed that...er...more data could then be sent.
So, this was made part of flushing the channel.

An explanation of this bahvaiour has also been added to the flush function